Opinions of 3 AI models - March 2025 - Warning
Using  ChaGPT, DeepSeek and Perplexity with the same prompt... 

At every level of Ai i have seen this repeated lack of awareness of the depth of the topic, when doing programming examples, which makes me deeply doubt the use of AI for language work without a lot of probing. AI seems to vomit a set of information but has no proper framework to concisely study an issue.